pipeline/requirements.txt
pybaseball
supabase
pandas
python-dotenv
apscheduler
